Update Bayes estimation to accept adjustable priors
Adjusted several classes within the Bayes estimation process to be able to accept an adjustable prior parameter. This allows for more flexibility when modeling and estimating due to the parameter's ability to handle different prior probabilities. Additionally, refactored some variable names to better reflect their functions.

+++ tetrad-lib/src/main/java/edu/cmu/tetrad/bayes/CptMapCounts.java
@@ -175,7 +175,7 @@ public class CptMapCounts implements CptMap {
     /**
      * The prior count for all cells.
      */
-    public void setPriorCount(int priorCount) {
+    public void setPriorCount(double priorCount) {
         this.priorCount = priorCount;
     }
 }
